Structural changes. Reorganized samples and exports
