+++ demos/app/app.module.ts
@@ -2,21 +2,15 @@ import { NgModule } from "@angular/core";
 import { BrowserModule } from "@angular/platform-browser";
 import { FormsModule } from "@angular/forms";
 
-import {
-
-    IgInput, CarouselModule , CheckboxModule, SwitchModule,
-    IgRadioModule , TabBarModule , ListModule, ButtonModule, Label,
-    AvatarModule, FilterModule, IgRippleModule, BadgeModule, IgxIconModule
-
-} from "../../src/main";
+import { IgxComponentsModule, IgxDirectivesModule } from "../../src/main";
 
 import { AppComponent } from "./app.component";
-import { InputSampleComponent } from "./inputsample.component";
-import { CarouselSampleComponent } from "./carouselsample.component";
-import { TabBarSampleComponent } from "./tabbarsample.component";
-import { ListSampleComponent } from "./listsample.component";
-import { ButtonsSampleComponent } from "./buttonssample.component";
-import { AvatarSampleComponent } from "./avatarsample.component";
+import { InputSampleModule } from "./input/sample.module";
+import { CarouselSampleModule } from "./carousel/sample.module";
+import { TabBarSampleModule } from "./tabbar/sample.module";
+import { ListSampleModule } from "./list/sample.module";
+import { ButtonSampleModule } from "./button/sample.module";
+import { AvatarSampleModule } from "./avatar/sample.module";
 import { NavDrawerSampleModule } from "./navigation-drawer/sample.module";
 import { NavbarSampleModule } from "./navbar/sample.module";
 import { ModalSampleModule } from "./modal/sample.module";
@@ -27,34 +21,22 @@ import { routing } from "./app.routing";
     imports: [
         BrowserModule,
         FormsModule,
-        CarouselModule,
-        CheckboxModule,
-        SwitchModule,
-        IgRadioModule,
-        TabBarModule,
-        ListModule,
-        ButtonModule,
-        AvatarModule,
-        FilterModule,
+        routing,
+        IgxComponentsModule,
+        IgxDirectivesModule,
+        InputSampleModule,
+        CarouselSampleModule,
+        TabBarSampleModule,
+        ListSampleModule,
+        ButtonSampleModule,
+        AvatarSampleModule,
         NavDrawerSampleModule,
         NavbarSampleModule,
         ModalSampleModule,
-        IconSampleModule,
-        IgRippleModule,
-        routing,
-        BadgeModule,        
-        IgInput,
-        IgxIconModule
+        IconSampleModule
     ],
     declarations: [
         AppComponent,
-        InputSampleComponent,
-        CarouselSampleComponent,
-        TabBarSampleComponent,
-        ListSampleComponent,
-        Label,
-        ButtonsSampleComponent,
-        AvatarSampleComponent
     ],
     bootstrap: [
         AppComponent
