+++ demos/app/avatar/sample.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from "@angular/core";
+import { AvatarSampleComponent } from "./sample.component";
+import { IgxComponentsModule, IgxDirectivesModule } from "../../../src/main";
+
+@NgModule({
+    imports: [IgxComponentsModule, IgxDirectivesModule],
+    declarations: [AvatarSampleComponent]
+})
+export class AvatarSampleModule {}
