+++ demos/app/button/sample.component.ts
@@ -1,0 +1,9 @@
+import { Component } from "@angular/core";
+import { ButtonModule } from "../../../src/button/button";
+
+@Component({
+    selector: "button-sample",
+    moduleId: module.id,
+    templateUrl: './sample.component.html'
+})
+export class ButtonsSampleComponent { }
