+++ demos/app/button/sample.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from "@angular/core";
+import { IgxDirectivesModule } from "../../../src/main";
+import { ButtonsSampleComponent } from "./sample.component";
+
+@NgModule({
+    imports: [IgxDirectivesModule],
+    declarations: [ButtonsSampleComponent]
+})
+export class ButtonSampleModule {}
