+++ demos/app/carousel/sample.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import { IgxComponentsModule } from "../../../src/main";
+import { CarouselSampleComponent } from "./sample.component";
+
+@NgModule({
+    imports: [IgxComponentsModule, CommonModule],
+    declarations: [CarouselSampleComponent]
+})
+export class CarouselSampleModule {}
