+++ demos/app/list/sample.component.html
@@ -1,0 +1,50 @@
+<div class="wrapper">
+    <h4>Data Source Filtered List</h4>
+    <div class="ig-form-group">
+        <input class="ig-form-group__input--search" placeholder="Search List" [(ngModel)]="search1" />
+        <label igLabel>
+            <i class="material-icons">search</i>
+        </label>
+        <span class="ig-form-group__clear--hidden">
+            <i class="material-icons">clear</i>
+        </span>
+    </div>
+    <ig-list>
+        <ig-list-item igRipple="pink" igRippleTarget=".ig-list__item" *ngFor="let item of navItems | filter: fo1">
+            {{item.text}}
+        </ig-list-item>
+    </ig-list>
+</div>
+<div class="wrapper">
+    <h4>Declarative Fitered List</h4>
+    <ig-checkbox [checked]="true" #checkbox>Perform filtering</ig-checkbox>
+    <div class="ig-form-group">
+        <input class="ig-form-group__input--search" placeholder="Search List" [(ngModel)]="search2" />
+        <label igLabel>
+            <i class="material-icons">search</i>
+        </label>
+        <span class="ig-form-group__clear--hidden">
+            <i class="material-icons">clear</i>
+        </span>
+    </div>
+    <ig-list #declarativeList [filter]="fo2" (filtering)="filteringHandler($event)" (filtered)="filteredHandler($event)">
+        <ig-list-header>Mildly Sweet</ig-list-header>
+        <ig-list-item [options]="{}">Red Delicious</ig-list-item>
+        <ig-list-item>Ambrosia</ig-list-item>
+        <ig-list-item>Rome</ig-list-item>
+        <ig-list-header>Sweet</ig-list-header>
+        <ig-list-item>Golden Delicious</ig-list-item>
+        <ig-list-item>Cosmic Crisp</ig-list-item>
+        <ig-list-item>Pinova</ig-list-item>
+    </ig-list>
+</div>
+<div class="wrapper">
+    <h4>Non-header List with options</h4>
+    <ig-list>
+    <ig-list-item *ngFor="let navItem of navItems; let index = index" [options]="options">
+        <div>
+            {{navItem.text}}
+        </div>
+    </ig-list-item>
+    </ig-list>
+</div>
