+++ demos/app/list/sample.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import { FormsModule } from "@angular/forms";
+import { IgxComponentsModule, IgxDirectivesModule } from "../../../src/main";
+import { ListSampleComponent } from "./sample.component";
+
+@NgModule({
+    imports: [IgxComponentsModule, IgxDirectivesModule, CommonModule, FormsModule],
+    declarations: [ListSampleComponent]
+})
+export class ListSampleModule {}
