+++ demos/app/tabbar/sample.component.ts
@@ -1,0 +1,9 @@
+import { Component } from "@angular/core";
+import { IgxComponentsModule, IgxDirectivesModule } from "../../../src/main";
+
+@Component({
+    selector: "tabbar-sample",
+    moduleId: module.id,
+    templateUrl: './sample.component.html'
+})
+export class TabBarSampleComponent { }
