+++ demos/app/tabbar/sample.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from "@angular/core";
+import { IgxComponentsModule, IgxDirectivesModule } from "../../../src/main";
+import { TabBarSampleComponent } from "./sample.component";
+
+@NgModule({
+    imports: [IgxComponentsModule, IgxDirectivesModule],
+    declarations: [TabBarSampleComponent]
+})
+export class TabBarSampleModule {}
